moved from culling to geospatial
